move ui to a crate
