+++ crates/shared/src/lib.rs
