+++ crates/shared/src/window_handle_hack.rs
