+++ crates/ui/src/app/icons.rs
