+++ crates/ui/src/app/music_cache.rs
