+++ src/main.rs
@@ -5,9 +5,9 @@ use clef_audio::player::{AudioAction, AudioMessage, Player};
 use clef_db as db;
 use clef_db::run_migrations;
 
-use clef::app::config::Config;
-use clef::app::{App, Flags};
 use clef::logging;
+use clef_ui::app::config::Config;
+use clef_ui::app::{App, Flags};
 
 fn main() -> iced::Result {
     logging::init();
